backend/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.7
pytest==7.4.4
httpx==0.26.0
tenacity==8.2.3
python-dotenv==1.0.0 